Gets collisionDetectCylinder test to work.
Had to increase the nymber of vertices and iterations for EPA *a lot*:
  * EPA_MAX_VERTICES:   128 -> 4096
  * EPA_MAX_ITERATIONS: 255 -> 4096
I'm not comfortable increasing them this much for everything - the decrease in
speed for EPA is quite noticable. Ideally, I'd pass them in on constructing the
penetration depth solver for a collision pair involving cylinders. This would
require re-working `btGjkEpaPenetrationDepthSolver` and `btGjkEpa2` to change
these settings from pound defined values to settable values.